simplified package structure, file rename and caught a bug where sync crashes if ignores were not defined
